requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cleanup==7.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
django-tinymce==3.5.0
django-widget-tweaks==1.4.12
Pillow==9.4.0
python-dotenv==1.0.0
six==1.16.0
sqlparse==0.4.3
